chore: Moves storage-pointer one level up
